Move YARDs extensions to Gem::Specification and Gem::DocManager into 'yard/rubygems'.
* Since RubyGems 1.4 disabled RubyGems Plugin loading for non-gem
  commands, YARDs extensions to RubyGems cannot be loaded by normal Ruby
  code. Moving these extensions into 'lib/yard', allows 'require' to
  load them again.
